CSCI5117 Interactive Web/homework-1-allen-liaoo/src/requirements.txt
-i https://pypi.org/simple
blinker==1.7.0 ; python_version >= '3.8'
click==8.1.7 ; python_version >= '3.7'
flask==3.0.2
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.3 ; python_version >= '3.7'
markupsafe==2.1.5 ; python_version >= '3.7'
psycopg2-binary==2.9.9
werkzeug==3.0.1 ; python_version >= '3.8'